fix: build fixes for ttsim parity after multichip merge
Add memory_pins to mesh command queue memcpy submit APIs, fix Tracy
marker color for the pinned submodule, and bump UMD with sim build fixes.

+++ tt_metal/impl/dispatch/realtime_profiler_tracy_handler.cpp
@@ -35,7 +35,7 @@ tracy::TTDeviceMarker make_marker(
     marker.marker_type = type;
     marker.file = file_str;
     marker.line = 0;
-    marker.color = tracy::Color::LightBlue;
+    marker.color = static_cast<uint32_t>(tracy::Color::LightBlue);
     return marker;
 }
 #endif
